+++ test.sh
@@ -7,10 +7,10 @@ die() { set +v; echo "$*" 1>&2 ; sleep 1; exit 1; }
 
 server_up() {
   TRIES=0
-  MAX_TRIES=10
+  MAX_TRIES=60
   URL=http://localhost:$1
   until curl --silent --fail $URL; do
-    [ ${TRIES} -gt ${MAX_TRIES} ] && die 'Server did not come up'
+    [ ${TRIES} -gt ${MAX_TRIES} ] && die "Server not running at $URL"
     printf '.'
     sleep 1
     TRIES=$(($TRIES+1))
@@ -39,6 +39,7 @@ end pytest
 start docker
 ./docker.sh 5001
 server_up 5001
+npm run cypress:run
 end docker
 
 start changelog